Progress on weight model
