+++ examples/weight_tester/producer_1
@@ -1,0 +1,1 @@
+java -cp .:$PILOTS_HOME/lib/* pilots.util.FileInputProducer ../../data/WeightTester/Normal.txt
