+++ examples/weight_tester/w_trainer.trn
@@ -1,0 +1,9 @@
+trainer w_model_trainer;
+  data
+    time, v_a, prs, tmp, alt_i, curr_w using file("weight_training.csv");
+  model
+    features: time, v_a, prs, tmp, alt_i;
+    labels: curr_w;
+    algorithm:
+      weight;
+end;
